requirements.txt
altair==4.2.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cmake==3.22.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.4
dlib==19.24.0
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
gcloud==0.17.0
gitdb==4.0.9
GitPython==3.1.27
googleapis-common-protos==1.56.1
httplib2==0.20.4
idna==3.3
image==1.5.33
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
jws==0.1.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
MultiApp==1.0
mysql-connector-python==8.0.15
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
oauth2client==3.0.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.4.3
pydeck==0.7.1
Pygments==2.12.0
Pympler==1.0.1
PyMySQL==1.0.2
pyparsing==3.0.9
Pyrebase==3.0.27
pyrsistent==0.18.1
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==22.3.0
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.8
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
streamlit==1.9.0
terminado==0.13.3
tinycss2==1.1.1
tk==0.1.0
toml==0.10.2
tkVideo==0.1
toolz==0.11.2
tornado==6.1
traitlets==5.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
validators==0.19.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
zipp==3.8.0